Wind：修改模型

+++ lib/views/charts/radio_gauge_chart/model/radio_gauge_chart_model.dart
@@ -1,10 +1,13 @@
+import 'dart:math';
+
+import 'package:decimal/decimal.dart';
 
 import '../../../../public.dart';
 
 class RadioGaugeChartData {
   final String? title;
-  final int? current;
-  final int? all;
+  final double? current;
+  final double? all;
   final Color? color;
   final String? icon;
   final String? symbol;
